@@ -9,8 +9,9 @@
   },
   "dependencies": {
     "98.css": "^0.1.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "motion": "^12.23.24",
+    "react": "^19.2.0",
+    "react-dom": "^19.2.0"
   },
   "devDependencies": {
     "@vitejs/plugin-react": "^2.2.0",
